Inspector : Broaden criteria for non-editable plug
Plugs can be non-editable for more reasons than being read-only :
- If a plug has an input
- If a plug is driven by the default row of a spreadsheet, we want it
to not be editable because that could affect more than the user expects.

If a plug is animated, the read-only and has-input tests are applied to
the `CurvePlug` that is the input to the plug in question.

These conditions apply to both single value plugs and compound plugs.
For compound plugs, the conditions apply to the compound plug itself as
well as each child plug, recursively. If any plug meets these
non-editable conditions, the compound plug is considered to be
non-editable.