Merge : Handle exceptional values in Difference mode
We want "inf" and "inf" to have a difference of 0, since they are colloquially the same. More concerningly,
we want NaN to return a difference of inf to anything other than NaN - previously, assertImagesEqual would pass
an image with NaN pixels, since the difference between NaN and any value is NaN, and NaN is not greater than a specified tolerance.