Cycles GeometryAlgo : Convert uniform normals to standard Cycles attribute
This prevents them from getting clobbered completely by Cycles' generation of default normals if it doesn't find the standard ones. As shown by the unit tests, our custom normal makes it through and is now queryable with an `attribute` node. It doesn't seem to actually affect the normal used by Arnold for shading though, but as far as I can tell, that is out of our hands.