LightEditor : Fig toggle scope
Previously, we were not scoping the viewer context when attempting to
toggle a boolean value automatically. Because the `__toggleBoolean()`
method uses `fullAttributes()` to determine the new toggle value, it
would not take the script context into account. This in turn caused
things like the mute toggle to not work if one of a light's attributes
did not provide a default value in a script context variable expression.