ParameterInspector : Light `OptionalValuePlug`
Two things are needed to make `OptionalValuePlug`, used by USDLuxLights,
compatible with `ParameterInspector` and `LightTool` :

1. Recognize when the `source` is coming from an `OptionalValuePlug`
and only return it as the source when the plug is enabled.
2. Recognize when the parameter does not exist in the shader data and
don't return an `editFunction()` if so.

Previously we were not checking if a parameter existed for a shader, so
we would return an `editFunction` that when called, would raise an
exception when it would be called via `acquireEdit()`.
( `::parameterValue` in `EditScopeAlgo` raises the exception. )