SpotLightHandle : Highlight only in highlight state
I'm not sure why we were highlighting also if the drag was active, all of my testing indicates the highlight state is sufficient to keep the highlight active while dragging.

This hasn't caused problems for the `SpotLightHandle`, but it caused multiple `QuadLightHandle` to be highlighted incorrectly, so I'm matching the behavior here for consistency.