TestRunner : Generalise test categories
This builds on the work done in #5358, allowing tests to be assigned to arbitrary categories using a new `CategorisedTestMethod` decorator. It also offers improved filtering by category in `gaffer test` by supporting wildcards and adding a `-excludedCategories` argument.

This will allow us to use test categories to group tests that are related in some way and are commonly run together during development.