3Delight : Refactor array parameter conversion
3Delight loads parameter data as a flat array, requiring extra metadata
on the parameter to express the variety of data we need. The `count`
metadata is used to express that there are multiple data entries for the
parameter, such as when storing primitive variables.

The `NSIParamIsArray` flag indicates the data entries are arrays and the
`arraylength` metadata indicates the length of each array.

Setting the `count` to the length of the array is appropriate for
primitive variables. But doing so in `ParameterList::parameter`, as we
were previously, prevents us from loading arrays for things like spline
shaders without additional ABI-breaking API.

Translating Gaffer's `*VectorData` as a single entry array by default
and modifying it in the case of primitive variables in `CurvesAlgo`,
`MeshAlgo` and `NodeAlgo` supports all data needs without additonal API.