Dispatch App : Fix setting of dispatch plug from command line
The same dispatcher can be added to the `dispatchers` list multiple
times, once from the `dispatcher` parameter, and again from the
`alternateDispatchers` parameter.

That in turn resulted in the wrong dispatcher object being the one
updated by the settings from the command line.

The fix here is to prevent the duplicate addition of the same
dispatcher.