Dispatcher : Remove `createMatching()` static method
This was a premature addition to the API - to our knowledge it only ever had one use case, and this eventually turned out to be poorly served by it anyway - see #5434. It's better to conform to the more minimal registry API used by other classes, and then let people compose their own logic on top.