SceneAlgo : Add `findAllWithAttribute()` method
I did spend a while thinking about the possibility of promoting the existing private `findAttributes()` function
from SceneAlgo.cpp instead. But that's a fairly esoteric thing that runs a predicate functor on the full attributes, and for `findAllWithAttribute()` my goal was to provide something simpler which could be used from Python without the overhead of a Python functor.