ImageSampler : More thorough tests, more accurate bound
When looking at this code, it was weird that the sampleWindow used a simple conversion from float to int, which roundts towards 0, meaning that the alignment of the actual sampling to the sampleWindow was different for pixels with negative coordinates. I thought this might be producing bugs - but turns out that Sampler expands the sampleWindow anyway, so the input only needs to contain the one pixel you are closest to. I added some tests that test negative pixels and different offsets within the pixel, and then changed ImageSampler to always use a minimum sampleWindow with a consistent alignment to the pixel.