RenderController : Tweak camera shutter change tracking
The dirty flags track things that may no longer be in sync with the Gaffer scene, and therefore require an update. The changed flags track things that actually changed when we performed that update. Use this consistently for camera shutter updates, and improve comment to clarify why we're doing what we're doing.

As far as I can tell, this doesn't affect the correctness of any outcome, but it does mean we'll avoid some redundant work when restarting after a cancellation part way through an update. This is because we clear the dirty flags as soon as we've updated the globals, but don't clear the changed flags until we've updated the whole scene.