RendererAlgo : Consolidate options handling
Both RenderController and LocationOutput had a little class for dealing with the most important rendering options. Consolidating them into a single shared RenderOptions class lets us remove a bit of duplicate code and simplify a few function signatures. But the main purpose of this rejig is to set us up so that we can pass the RenderOptions to Capsules so that they respect the correct motion blur settings.