RenderController : Remove redundant conditional
We know that type isn't LightType or LightFilter type already, because there is a return statement for that case at line 856.

I've also replaced the non-const `isNull` variable with a test directly inside the `if`, so there's one less piece of state to have to bear in mind throughout the rest of the function (where the variable isn't actually used again).