FileSystemPath : Work around Windows exFAT bug
MSVC's `std::filesystem` has a bug that causes it to fail to retrieve
file information on exFAT partitions in versions prior to (roughly)
17.2. That version is the first release after merging a fix from
https://github.com/microsoft/STL/pull/2373, but I'm not certain that
version has the fix included.

Using `std::filesystem::status` instead of `symlink_status()` seems
justified because exFAT paritions don't support symlinks. If other
partition types are included in the filter for error code 87, this
may need to be revisited.