PythonEditor : Fix lifetime management bug
Because MessageHandler is a wrapped class, it only gets destroyed during garbage collection. This extended the life of the PythonEditor unnecessarily, which is the sort of thing that causes the classic PySide "C++ object already deleted" error. Although we haven't seen those for the PythonEditor yet, its worth fixing this before we do.