_LinkedScrollBar : Guard against recursive updates
This may fix #5296 - though I haven't been able to reproduce it. The
problem appears to be endless recursion of `sliderChange()`, which may
have been caused by one of the two scroll bars the `LinkedScrollBar`
manages emitting a change signal that then causes the other scroll bar
to update, emitting a change signal... until the stack overflows.