Reference : Remove child edits from `m_plugEdits`
Previously, we were only removing edit entries for the immediate
children of a Reference. This could leave stale edits in the edit
cache for plugs at the grandchild level and deeper. Usually that does
not cause a problem because `m_plugEdits` is keyed using the plug
pointer address. But on Windows in particular, the address of a
destroyed plug was occasionally (something like 1 in 1000) being reused
for a new plug, causing a false positive lookup in `m_plugEdits`.