gui app : Fix copy and paste on Windows
It seems that Windows or Qt on Windows emits an extra update signal
when copying to the clipboard, thwarting our attempt at not receiving
the Qt clipboard update event when syncing the Gaffer clipboard. By
checking to make sure there is a real change to make, we don't clobber
the Gaffer clipboard's stored `IECore::Object` with a string.

Fixes #5485