CollectImages : Introduce MappingData
This removes quadratic performance in number of layer names, because we were calling `sourceLayerAndChannel()` once per output channel, and it was looping over all the input layers. This gives a modest 30% reduction in runtime for `testHighLayerCountPerformance`. That's not the main motivation for MappingData though - it provides the necessary foundations for the next commit.