ParallelAlgo : Don't throw for missing UIThreadCallHandler
Instead output a diagnostic message.

This isn't something we expect to ever happen in normal GUI usage, because we push a single call handler and leave it there forever. But in the unit tests, we regularly push and pop handlers using the `ParallelAlgoTest.UIThreadCallHandler` context manager. Here, throwing was incredibly unhelpful, due to the following sequence :

1. Enter a UIThreadCallHandler.
2. Do something that will cause calls to be made on a background thread.
3. Assert something that fails, raising an exception and exiting the UIThreadCallHandler.
4. Background thread call continues and finds that there is no handler, and throws, causing process termination due to unhandled an exception.
5. We never get to see the assertion failure from 3, which is the key piece of information we need.