Catalogue : Improve recognition of displays from the same render
The change to `testDisplayDriverAOVGrouping` is necessary because it was exploiting a bug in the previous implementation. As soon as the first image was sent the Catalogue was saving the image (because the driver was closed), but then still allowing other drivers to be added to the image (causing another save after that). We now reject new drivers as soon as we start a save, so the test needs to keep the first driver open until the second has been created. The test now matches more closely the sequence of events we'd see in a real render.