Dispatcher : Rederive from TaskNode
This allows Dispatchers to be composed in a task graph to form jobs comprising multiple dispatchers. For example :

- A LocalDispatcher with a Wedge, launching multiple TractorDispatcher jobs to a farm.
- A LocalDispatcher upstream of a TractorDispatcher, to launch a subset of tasks as a single Gaffer process on a single machine.

The `dispatch()` method has been removed, since `execute()` is the standard way of invoking a TaskNode, and the `tasks` plug defines the tasks to be executed. But a compatibility config monkey-patches `dispatch()` back in, implemented in terms of `execute()`. This keeps all our unit tests running, and will support existing pipeline scripts depending on the old API.