DispatcherUI : Remove "Execute" menu items and buttons
Now that Dispatchers are TaskNodes, we want people to host them in the node graph itself, so we are removing the global dispatchers hosted in the UI. This means removing the "Execute" button on each individual TaskNode, but we've replaced it with a "Create a Dispatcher" helper widget which guides the user into the new workflow. A new "Dispatch" button on the dispatcher itself is now the only way to dispatch tasks.