LocalDispatcher : Simplify status tracking
We were storing a status per batch, but in fact the only status we ever
reported to the outside world (via `failed()` and `killed()`) was the
status of the root batch. And to get the root batch into a failed state
we were going via a convoluted route whereby `__reportFailed()` would
mark the current batch as failed, and call `JobPool._failed()` which would
in turn call back to `Job._fail()` to mark the root batch as failed. Much
easier to manage a single `self.__status` directly, and limit the per-batch
data to a simple flag that tells us if we've executed it yet or not.

At the same time, we explicitly store the current batch instead of having
to do a search for it every time. Note that we're assuming that attribute assignment (for `self.__currentBatch`) is atomic (it is for CPython) rather than using a mutex to synchronize access between the background and UI threads.