LocalDispatcher : Keep all jobs in pool until explicitly removed
Otherwise you get no opportunity to see the job logs in the LocalJobs
panel, and for quick jobs its not even clear if the job happened. Also
add a `Job.status()` method since there was no public way of determining
running/complete status before (it was assumed that anything in the list
that wasn't killed or failed was running).

Remove `Job.killed()` and `Job.failed()` since they are superceded by
`Job.status()`.