ArnoldShader : Support `camera_projection.camera` parameter
This is just a case of setting the `gaffer.plugType` to "StringPlug" and letting the recently added `NodeParameter` mechanism do the rest in `ArnoldRenderer`.

Also added a nice camera browser using ScenePathPlugValueWidget. I was unsure what to do for this in ArnoldShaderUI, as the "camera" value for "widget" isn't part of the OSL standard. The other option would have been to support three new Gaffer-specific metadata entries for `gaffer.plugValueWidget.type`, `scenePathPlugValueWidget.setNames` and `scenePathPlugValueWidget:setsLabel`. But that seemed quite faffy when we don't have any other use cases for them at present, and "camera" seems like something that could reasonably be supported by other DCCs.