IECoreArnold : Support `NODE *` shader parameters
We do this by accepting StringData parameters with the path to a node, and then resolving that to a `NODE *` at the very last minute in `ArnoldRenderer::render()`. This is robust to the targeted node being deleted or replaced. In an ideal world, Arnold's `camera_projection` shader would just have a string parameter instead, like its `toon.lights` parameter.

We do the resolution to `NODE *` in our ShaderCache, which means there are currently a couple of caveats :

- Shaders inside procedurals are not supported, because ProceduralRenderer uses its own ShaderCache which is destroyed after expansion. We do want procedurals to share the main caches in future anyway, at which point this would "just work".
- Light shaders are not supported, because they're not stored in the cache because lights need unique ownership. Hopefully this isn't a likely use case. One potential solution might be to have ArnoldLight pull from the ShaderCache, but passing an argument that always uniquefies the hash.