Catalogue : Fix duplicate save processes when IPR ends
This bug was introduced in d67ace35f72dbc12d98156f6cd2f8c54bf8ccc54. What I'd failed to realise at the time was that `Display::driverClosed()` will return `True` _before_ `Display::imageReceivedSignal()` has been emitted on the UI thread, because it's just an accessor to an atomic flag. This means that its the _first_ call to `InternalImage::driverClosed()` that will initiate the saving process, rather than the last call as before. So in subsequent calls we need to check we haven't yet initialised `m_saver` to avoid repeating the save process unnecessarily.

For people with a lot of AOVs, the slowdown from the multiple saves could be quite drastic. What made it even worse was that the assignment of `m_saver = newRedundantAsynchronousSaver` would block, because the destructor for the old saver waits for the background process to finish.