BackgroundMethodTest : Increase `waitForIdle()` count
We've been seeing regular failures on Windows CI in `assertIsNone( w() )`
at L259. `w` can only die when the BackgroundMethod has returned, so my
assumption is that when a CI machine is under heavy load, we're not
waiting long enough for this to happen. So I'm upping the limit in the
hope that it helps matters.