GraphEditor : Fix drag handling errors
We can't assume that if `dropSignal()` is emitted it was because _we_
accepted `dragEnterSignal()`. Another slot connected to `dragEnterSignal()`
may have accepted it instead, in which case we don't want to do anything in
our `GraphEditor.__drop`, on the assumption that another connection to
`dropSignal()` will be dealing with it.