Dispatch : Avoid quadratic scaling in `Batcher::addPreTask()`
This actually didn't scale too badly for the frame ranges we typically use,
but it dominated dispatch time for large frame ranges like the one used in `DispatcheTest.testBasicDispatcherPerformance`. Fixing this knocks more than 60% off the total runtime of that test.

This is another reason for making `TaskBatch::preTasks()` const-only in a previous commit, as it means we can maintain sychronisation between `m_preTasks` and `m_preTasksSet` without fear of it being thrown off by edits from the outside.