Dispatcher : Make TaskBatch constructors private
They should only be created by the Batcher for consumption by Dispatcher subclasses, and it wasn't possible to make a _useful_ one even with the public constructors, because you couldn't add frames or pretasks (by design).

Also removed the deprecated constructor, and implemented the default constructor to forward to the standard one.