DispatcherTest : Add basic performance test
This is particularly uncharitable to the current Dispatcher implementation, because it uses a long frame range that is unlikely to be used in the real world. The current Dispatcher implementation scales poorly for that number of frames, but we're about to fix that, and then the long frame range will give  us a test duration in roughly the ballpark we want.