TaskNode : Remove deprecated `Task( taskNode, context )` constructor
In favour of the constructor taking a TaskPlug. As with elsewhere in Gaffer, plugs are the API via which the internal functionality of nodes is accessed, and callers shouldn't care what node they are on. The remaining constructors are easier to use in `TaskNode.preTasks()` because they make it clear that you're not responsible for following input connections to find a node, and can instead just return local plugs.