LocalJobs/LocalDispatcher : Improve memory/cpu update
This was far too slow when multiple jobs were in flight, and was causing the UI to stutter. It also didn't work on Windows.

There were a couple of obvious problems :

1. `ps -A` lists _all_ processes, so we had to sift through increasing amounts of data to find what we wanted. I think this would have been fixable with `ps -p <PID>`.
2. Both the CPU and Memory columns call `statistics()`, so we do twice the work we want to.

I'm not sure either of those were severe enough to be the cause of the stutter though, and I suspect the bigger problem was the launching of a subprocess.

One solution I tested was to read the memory information from `/proc/<PID>` (which is apparently where `ps` gets it from in the first place). This did fix the performance problem, but wouldn't have worked on MacOS where `procfs` isn't included. It would also have required more code to get CPU usage. The `psutil` module does this all for us, and more besides on all platforms we care about, and seems worthy of adoption as a dependency.