TransformTool : Support editing of XYZ component source inputs
Until now, we were assuming that the plugs passed to `canSetValueOrAddKey()`
did not have inputs, because we had already called `spreadsheetAwareSource()`
to find the source plug. But that operates on V3fPlugs, and is blind to
individual connections to each XYZ component (recall that the parent plug only
has an input connection if all the children have connections from the children
of another parent). This meant that if you promoted just one component to a box,
or promoted several components individually, the transform handles would be
disabled. The solution is to call `source()` one more time to find the plug we actually
want to edit.

At one point, I thought that `canSetValueOrAddKey()` was destined for more
general usage, possibly in PlugAlgo. But the function in its new form isn't
suitable for that at all, as it is tailored to the specifics of TransformTool
operation. It would have been possible to keep it in its "pure" form, but only
at the expense of duplicating the new checks into each of the derived classes.