ScaleTool : Deal with case where all components are driven by one plug
For instance, if every component had a connection from the same promoted
`uniformScale` plug. The problem here was that we were calling `setValueOrAddKey()`
for all the components, even the ones not affected by the individual handle being
dragged. Since each component is driven by the same plug, we'd be setting that
plug three times, and only the Z component would win because it was set last. But
if the Z component wasn't affected by the current handle, it would be setting the
scale back to its _original_ value. We now consider the plugs that are affected
by the current handle to avoid this clobbering. In the `uniformScale` case this
has the expected effect - no matter which handle you pick, you get uniform scaling.