GafferScene : Register standardOptions
This is not an exhaustive registration of standard options, just a demonstrative set of the ones we require for the currently registered Render Pass Editor columns.

The default value we register for `option:render:resolution` matches the default we display in the Node Editor for the StandardOptions node's `renderResolution` plug. As far as a render is concerned the default comes from IECoreScene::Camera's default values for resolution, but the StandardOptions default is a better starting point.