GafferImageTest : Add deep test image data. These images are ...
These images are intended to make it easy to exercise all cases in deep processing code.

Color values are uniform random, alpha values are a mixture of uniform random values, but also: values clustered very near 1, values of 1, and values of 0 ( these are the values most likely to trigger special cases in deep processing )

Depth values range from 0 to 4.  The difference in the 4 images are how depth values are generated. "Points" means Z and ZBack are set the same, whereas "Volumes" means that ZBack is allowed to range higher than Z. "Float" means that the depths are uniformly distributed, whereas "Int" means depths are limited to integers ( making it more likely that doing a deep merge will produce exact collisions ).

I've put the script used to generate these images in contrib - we could save 170K of test file size by generating these on the fly, but we try to avoid having a dependency on OSL in other tests, and having constant test images makes sense.