LightPositionTool : Persist handle across scene transforms
Previously we were using the final scene transform in world space to
store and calculate the pivot and target points. This meant when
switching from a node upstream of a `Group` node with a transformation,
we would lose the handle because the world space position would not
align with the handle.

Resetting in that case is not necessary if we instead store the
world space transformation from the scene that will receive the edits,
what the `TransformTool` calls the transform space.

Also, by using the upstream path as the key into the pivot and target
maps, we can persist the handle across hierarchy changes if the
transform allows.