SConstruct : Export node definitions using ExtensionAlgo
And convert Anaglyph node to demonstrate this in action.

There is a slight annoyance here in that the extensions depend on the whole of `buildCore`,
because we need to be able to run Gaffer to export them. Which means they get re-exported
when anything changes. This is pretty fast anyway, but it can be avoided around by using
`scons buildCore` rather than `scons build` when you know extensions don't need to be
exported again.