BackgroundTask/Editor : Coordinate to find cancellation subject
This hack allows us to find the cancellation subject for the internal plugs of the new ImageInspector. Although it's a tiny bit hacky, I'm reasonably content with it because it's pretty isolated and the new "Use Editor.Settings to store your settings" guidance moves us closer to where we want to be with regard to Editors being Nodes eventually. I suspect that when we get there we'll be able to replace it with a single cleaner mechanism for BackgroundTask to recover the ScriptNode from a UI element.