Arnold Renderer : Fix translation of `uchar` data from USD
Gaffer represents all integral data as IntData generated from IntPlugs, ignoring the fact that the underlying type in Arnold might be `AI_TYPE_BYTE`, `AI_TYPE_INT` or `AI_TYPE_UINT`. So the renderer backend was only looking for IntData in several cases, which was OK for data generated in Gaffer but not for data loaded directly from USD. We now check for the exact type in addition to checking for IntData.

Longer term it seems clear that we should change Gaffer to generate the exact datatypes as well. This probably means introducing new plug types, and changing the type of common plugs on ArnoldAttributes and ArnoldShader should probably be considered a breaking change, as should changing the data types being output. So that remains as future work targeted to a major version change.