Gaffer : Version the preferences location
This avoids problems caused by saving preferences from a later version in a format not compatible with an earlier version. The most common problem is saving a layout containing an editor that doesn't exist in the earlier version, and version 1.4 will shortly be introducing a new editor.

Fixes GafferHQ#368
Fixes GafferHQ#2882